Added sql() to pairing.php
I added the sql function to pairing.php. I kept the parameters and added a new one, pdo. The function should be sql injection protect and it should return a string.

Also added player and winner to the construct function. Thought it might be needed but I haven't checked out the requirements for the construct function yet so might modify it in that pull request.